+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Bugfix: 
   - incorrect table th sticky top when use multi header, because it use `props.rowHeight` to calculate top.(quick resolve: add `props.headerRowHeight` to fix).
   - TODO:  multi header fixed 
+  - fix th draggable always be true.
 
 ## 0.6.2
 * Bugfix: Click the sort triggered by the header. If the col.sorter is false, it will not be processed. Sorting triggered by setSorter remains popular.